Move auth functions to \'auth\' package
